+++ lib/brando/exceptions/registry_error.ex
@@ -1,6 +1,0 @@
-defmodule Brando.Exception.RegistryError do
-  @moduledoc """
-  Defines an exception for Registry errors.
-  """
-  defexception [:message]
-end