+++ test/test_helper.exs
@@ -249,5 +249,3 @@ Brando.Cache.Globals.set()
 Ecto.Adapters.SQL.Sandbox.mode(Repo, :manual)
 Brando.endpoint().start_link
 Brando.presence().start_link
-Brando.Registry.start_link()
-Brando.Registry.register(Brando, [:gettext])